Проведён рефакторинг 2
